Move feature checking to reconciler level
