+++ images/agent/src/internal/utils/block_device_test.go
@@ -52,7 +52,7 @@ var _ = Describe("BlockDevice", func() {
 				stat.Mode = S_IFBLK
 				return nil
 			})
-			sysCall.EXPECT().Syscall(uintptr(unix.SYS_IOCTL), fd, unix.BLKGETSIZE64, gomock.Any()).DoAndReturn(func(_, _, _, a3 uintptr) (uintptr, uintptr, Errno) {
+			sysCall.EXPECT().Syscall(uintptr(unix.SYS_IOCTL), fd, uintptr(unix.BLKGETSIZE64), gomock.Any()).DoAndReturn(func(_, _, _, a3 uintptr) (uintptr, uintptr, Errno) {
 				*(*uint64)(unsafe.Pointer(a3)) = uint64(size)
 				return 0, 0, 0
 			})
