--- images/agent/src/internal/utils/volume_cleanup_ce.go
@@ -1,27 +1,0 @@
-//go:build ce
-
-/*
-Copyright 2025 Flant JSC
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package utils
-
-import (
-	"context"
-	"fmt"
-
-	"agent/internal/logger"
-)
-
-func VolumeCleanup(_ context.Context, _ logger.Logger, _ BlockDeviceOpener, _, _, _ string, _ *RangeCover) error {
-	return fmt.Errorf("volume cleanup is not supported in your edition")
-}
